refactor: split demo components and rename actions
