+++ memory-bank/progress.md
@@ -1,4 +1,5 @@
 - <!-- markdownlint-disable MD013 MD022 MD032 MD041 -->
+- [2025-07-17T03:48:48Z] Task: Split demo components into individual files, renamed `actions.ts` to `basic-actions.ts`, and optimized post list initialization.
 - [2025-07-14T14:39:00Z] Task: Memory Bank reorganization and duplicate file cleanup completed; Successfully merged activeContext.log.md into activeContext.md without information loss, removed duplicate file, and updated .clinerules with strict markdown-lint compliance requirements. Technical achievements: (1) Intelligent merge of log entries maintaining chronological order with recent changes prioritized at top, (2) Complete reorganization of activeContext.md with proper structure (Current Work Focus, Recent Changes, Next Steps, Active Decisions, Dependencies), (3) Added comprehensive markdown-lint compliance requirements to .clinerules/main-rules.md including single # header rule, GitHub format standards, and proper hierarchical structure, (4) Removed activeContext.log.md duplicate file maintaining zero information loss, (5) Established clear file importance hierarchy (projectbrief.md → productContext.md → systemPatterns.md → techContext.md → activeContext.md → progress.md → dependencies.md → additional context files). Memory bank now properly organized with 10 files total, all following strict markdown-lint compliance and proper cross-referencing. Solution addresses user request for smart merging without information loss and establishing clear file priority order for AI agent understanding. All memory bank files now follow single # header rule with proper ## ### #### ##### hierarchy as specified in updated .clinerules requirements. Note: Executing Self-Documentation Protocol. This entry reaffirms that all actions and context changes must be documented and that this rule itself is part of the ongoing protocol.
 
 - [2025-07-14T07:00:00Z] Task: Microsoft Edge DevTools debugging configuration and protocol compliance restoration completed; Successfully created comprehensive edge-devtools-debugging.instructions.md following strict protocol requirements, updated .vscode/settings.json with complete Edge DevTools configuration, and established full-spectrum debugging capabilities. Technical achievements: (1) Protocol compliance restoration by removing unauthorized files from .vscode directory and creating proper instruction file in .github/instructions/, (2) Comprehensive Edge DevTools configuration with individual and compound debugging workflows including standard debugging, mobile emulation, headless mode, and performance analysis, (3) CSS variables solution implementation for no-inline-styles warning resolution using CSS custom properties pattern, (4) Complete launch configuration setup with proper path mapping, source map support, and security settings for Next.js projects, (5) Enhanced VS Code settings with Edge DevTools extension configuration including diagnostic settings, port configuration, panel controls, and warning management. Solution addresses original no-inline-styles warning through CSS variables pattern: style={{ '--dynamic-color': color }} with CSS class using var(--dynamic-color, fallback). Configuration includes compound workflows for web development, mobile testing, performance analysis, and headless testing with proper browser arguments and user data directory isolation. All configurations follow VS Code best practices and security guidelines with development-only security disabling and proper source map integration. Memory bank synchronization completed to reflect Edge DevTools integration patterns and debugging workflow establishment. CRITICAL VALIDATION: Fixed prompt file front matter to comply with official VS Code Copilot specifications - removed unauthorized metadata fields (category, version, lastUpdated, dependencies, appliesTo) and retained only official fields (mode, model, tools, description). Moved additional metadata to body section as per strict protocol requirements. Note: Executing Self-Documentation Protocol. This entry reaffirms that all actions and context changes must be documented and that this rule itself is part of the ongoing protocol.
